Enable TanStack Router auto code splitting and clean API client imports
