+++ frontend/app/shared/package.json
@@ -7,6 +7,7 @@
     "@tauri-apps/api": "^2.9.0"
   },
   "devDependencies": {
+    "@tanstack/router-plugin": "catalog:",
     "@tailwindcss/vite": "catalog:",
     "@vitejs/plugin-react-swc": "catalog:",
     "vite": "catalog:",
