--- frontend/core/src/scripts/api/httpClient.ts
@@ -1,2 +1,0 @@
-export { callApi, baseURL, httpClient, ApiError } from "./client";
-export type { CallApiOptions } from "./client";
