+++ frontend/core/src/scripts/api/services/DecoponSessionService.ts
@@ -5,7 +5,7 @@ import {
   DecoponSessionStatus,
 } from "@/scripts/types";
 import { endpoints } from "../endpoints";
-import { callApi } from "../httpClient";
+import { callApi } from "../client";
 
 export const DecoponSessionService = {
   progressPayload(): ApiRequestData {
