+++ frontend/core/src/scripts/api/services/ProfileService.ts
@@ -1,5 +1,5 @@
 import { endpoints } from "../endpoints";
-import { callApi } from "../httpClient";
+import { callApi } from "../client";
 import type {
   ApiRequestData,
   PreferenceResponse,
