+++ frontend/core/tsr.config.json
@@ -1,0 +1,5 @@
+{
+  "routesDirectory": "./src/routes",
+  "generatedRouteTree": "./src/routeTree.gen.ts",
+  "autoCodeSplitting": true
+}
