+++ package.json5
@@ -55,6 +55,7 @@
       axios: '^1.12.0',
       'brace-expansion': '2.0.2',
       devalue: '^5.4.2',
+      '@tanstack/router-generator': '1.120.10',
       'form-data': '^4.0.4',
       'vite@^5.0.0': '5.4.21',
       'vite@^6.0.0': '6.4.1',
